Adding Projectiles
Wow this is mor3e than I expected it would be.  Added a time util and
projectile util, and some stuff that might become some kind of parallel
execution utility.  Functions that run every tick until they don't.
Actually, that's just things that remove themselves from the time util,
huh?  What I have right now is just kind of a wrapper around that.
Guns, bombs, and pods have a good deal of flexibility, I'm toying with
adding more for angle and spread/formation for shots, but that seems
like it might be excessive.  ...But this is Jitsugyoka, and where Jitsu
is concerned, excess does not exist.